Add XtaRules02 and 03
I realised that it might be a problem that some of the words to be
replaced by the XtaRules begin with an apostrophe.
I made two more sets of XtaRules hoping that
one of them might help. It depends on the ordering of
the XtaRules and the ApostropheToQ rule:
XtaRules02:
If we want to apply the Xta rules before the ApostropheToQ rule.
However, I'm not sure if this
version would actually work.
XtaRules03:
If we want to apply the ApostropheToQ rule before the
XtaRules.