Add Xta and Apostrophe tests
I added the folder "xtaApostropheToQ". It contains a bunch of files for
testing the XtaRules and ApostropheToQ in one go.
The text.in-files are the same as for xta_simple and only the
text.out-files are different.
I thought it would be less confusing in the end, if I just create copies
of the text.in files and re-name them,
so now there are xtaApostropheToQ.text.in-files, too.
Let me know if you don't like the redundancy.
Since there is no apostrophe in <ngta> there's no combination test file
for it.

As the text.in files in this folder are the same as in
"xta_simple_tests" I did not make any addition to test_invariant in
testS.py.
And because I don't know about ordering rules yet, I haven't even tried
to add test_xtaToQ, so no testS7.py yet.